adding new fragments, refactoring packages
