--- app/src/main/java/com/example/sampleappbyme/main/fragment/SplashViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.sampleappbyme.main.fragment
-
-import androidx.lifecycle.ViewModel
-
-class SplashViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
