+++ app/src/main/java/com/example/sampleappbyme/main/view/fragment/MainFragment.kt
@@ -1,4 +1,4 @@
-package com.example.sampleappbyme.main.fragment
+package com.example.sampleappbyme.main.view.fragment
 
 import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
@@ -8,11 +8,13 @@ import android.view.View
 import android.view.ViewGroup
 
 import com.example.sampleappbyme.R
+import com.example.sampleappbyme.main.viewmodel.MainViewModel
 
 class MainFragment : Fragment() {
 
     companion object {
-        fun newInstance() = MainFragment()
+        fun newInstance() =
+            MainFragment()
     }
 
     private lateinit var viewModel: MainViewModel
@@ -27,6 +29,7 @@ class MainFragment : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         viewModel = ViewModelProviders.of(this).get(MainViewModel::class.java)
+
         // TODO: Use the ViewModel
     }
 }
