+++ app/src/main/java/com/example/sampleappbyme/main/view/fragment/RecyclerViewFragment.kt
@@ -1,0 +1,34 @@
+package com.example.sampleappbyme.main.view.fragment
+
+import androidx.lifecycle.ViewModelProviders
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import com.example.sampleappbyme.R
+import com.example.sampleappbyme.main.viewmodel.MainViewModel
+
+class RecyclerViewFragment : Fragment() {
+
+    companion object {
+        fun newInstance() =
+            RecyclerViewFragment()
+    }
+
+    private lateinit var viewModel: MainViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.recyclerview_fragment, container, false)
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        viewModel = ViewModelProviders.of(this).get(MainViewModel::class.java)
+        // TODO: Use the ViewModel
+    }
+}
