+++ app/src/main/java/com/example/sampleappbyme/main/viewmodel/MainViewModel.kt
@@ -1,7 +1,9 @@
-package com.example.sampleappbyme.main.fragment
+package com.example.sampleappbyme.main.viewmodel
 
 import androidx.lifecycle.ViewModel
 
 class MainViewModel : ViewModel() {
     // TODO: Implement the ViewModel
+
+
 }
